recipe_rag/requirements.txt
pandas
numpy
nltk
rouge-score
scikit-learn
sentence-transformers
transformers
torch
faiss-cpu
tqdm
